Update and rename README.txt to README.md
Statement of the problem

The prediction of oil and gas well production is one of the most important in the oil and gas industry. Without a reasonable production forecast, it is impossible to make decisions about the profitability of projects, capital investments, drilling new and operational planning for the operation of existing wells.

The purpose of building this  model is to predict one of the well performance parameters and to test the model's ability to predict the water cut of existing wells and wells that are planned to be drilled.

The production data is actually a time series, which requires building a more complex model. In order to simplify and speed up, the model is built for a specific date.